+++ src/extension.ts
@@ -7,10 +7,12 @@ import { registerLSP, deactivateLSP } from './lsp'
 import { registerBreakpoints } from './breakpoints'
 import { registerDocsProvider } from './bmxcommands'
 import { registerTaskProvider } from './taskprovider'
+import { registerBmxCliRequests } from './bmxrequest'
 import { registerHoverProvider } from './hoverprovider'
+import { registerColorProvider } from './colorprovider'
 import { registerBuildTreeProvider } from './buildtree'
-import { registerSymbolProvider } from './symbolProvider'
-import { registerFormatterProvider } from './formatterProvider'
+import { registerSymbolProvider } from './symbolprovider'
+import { registerFormatterProvider } from './formatterprovider'
 import { registerTerminalLinkProvider } from './linkprovider'
 import { registerCompletionProvider } from './completionprovider'
 import { registerSignatureHelpProvider } from './signaturehelpprovider'
@@ -23,6 +25,8 @@ export function activate( context: vscode.ExtensionContext ) {
 	registerDocsProvider( context )
 	registerTaskProvider( context )
 	registerHoverProvider( context )
+	registerColorProvider( context )
+	registerBmxCliRequests( context )
 	registerSymbolProvider( context )
 	registerBuildTreeProvider( context )
 	registerFormatterProvider( context )